+++ src/app/(main)/my-voyage/[teamId]/tech-stack/components/TechStackCard.tsx
@@ -16,7 +16,7 @@ import {
 import getTechCategory from "@/myVoyage/tech-stack/components/getTechCategory";
 import TextInput from "@/components/inputs/TextInput";
 import AvatarGroup from "@/components/avatar/AvatarGroup";
-import Avatar from "@/components/avatar/AvatarAlt";
+import Avatar from "@/components/avatar/Avatar";
 import Button from "@/components/Button";
 import type { TechStackItem } from "@/store/features/techStack/techStackSlice";
 import { useUser, useAppDispatch, useAppSelector } from "@/store/hooks";
@@ -230,7 +230,7 @@ export default function TechStackCard({ title, data }: TechStackCardProps) {
         </span>
       </div>
 
-      <div className="h-40 p-1 mt-6 overflow-y-auto">
+      <div className="mt-6 h-40 overflow-y-auto p-1">
         <ul className="text-base-300">
           {data.map((element) => {
             const voteIsSubmitted = element.teamTechStackItemVotes.find(
@@ -238,13 +238,13 @@ export default function TechStackCard({ title, data }: TechStackCardProps) {
             );
             return (
               <li
-                className="relative grid items-center grid-cols-6 mb-8 mr-2 text-base"
+                className="relative mb-8 mr-2 grid grid-cols-6 items-center text-base"
                 key={element.id}
               >
                 {editItemId === element.id ? (
                   <form
                     onSubmit={handleSubmitEdit(() => handleEdit(element.id))}
-                    className="h-12 col-span-6 -my-2"
+                    className="col-span-6 -my-2 h-12"
                     ref={editRef}
                   >
                     <TextInput
@@ -283,7 +283,7 @@ export default function TechStackCard({ title, data }: TechStackCardProps) {
                     {voteIsSubmitted ? (
                       <div className="relative col-span-2 flex w-[165px] items-center justify-end">
                         {element.teamTechStackItemVotes.length < 2 && (
-                          <div className="w-1/6 h-1/6">
+                          <div className="h-1/6 w-1/6">
                             <EllipsisVerticalIcon
                               className="mr-2 rounded-xl hover:cursor-pointer hover:bg-base-100"
                               onClick={() => setOpenMenuId(element.id)}
@@ -326,7 +326,7 @@ export default function TechStackCard({ title, data }: TechStackCardProps) {
         <div className="px-3.5 py-2.5">
           <Button
             variant="outline"
-            className="justify-center w-full border-2"
+            className="w-full justify-center border-2"
             onClick={toggleAddItemInput}
           >
             Add Tech Stack
