+++ src/app/(main)/my-voyage/[teamId]/voyage-resources/components/ResourceCard.tsx
@@ -2,7 +2,7 @@ import { ArrowTopRightOnSquareIcon } from "@heroicons/react/24/outline";
 import { TrashIcon } from "@heroicons/react/20/solid";
 import { useAppDispatch, useAppSelector } from "@/store/hooks";
 import { onOpenModal } from "@/store/features/modal/modalSlice";
-import Badge from "@/components/badge/BadgeAlt";
+import Badge from "@/components/badge/Badge";
 import IconButton from "@/components/IconButton";
 import { deleteResource } from "@/app/(main)/my-voyage/[teamId]/voyage-resources/resourcesService";
